+++ LawnCare/frontend/app/src/main/java/com/redflag/lawncare/common/api/ApiConstants.java
@@ -1,4 +1,4 @@
-package com.redflag.lawncare.common;
+package com.redflag.lawncare.common.api;
 
 public class ApiConstants {
     public static final String API_BASE = "https://ezmow-ca.myshopify.com/admin/api/";
